feat: add category page, add random products to main page
