+++ src/components/Aside.tsx
@@ -6,9 +6,8 @@ export const Aside = (): JSX.Element => {
   return (
     <aside className={styles.aside}>
       <Link to={PAGES.main.route}>Home</Link>
-      <Link to={PAGES.about.route}>About</Link>
+      <Link to={PAGES.catalog.route}>Catalog</Link>
       <Link to={PAGES.cart.route}>Cart</Link>
-      <Link to={PAGES.profile.route}>Profile</Link>
       <Link to={PAGES.signin.route}>Signin</Link>
       <Link to={PAGES.registration.route}>Registration</Link>
     </aside>
