+++ src/constants/common.ts
@@ -1,3 +1,4 @@
 export const CENTS_PER_EURO = 100;
 export const FIRST_INDEX = 0;
 export const PRODUCTS_IN_PAGE = 6;
+export const POPULAR_PRODUCTS_IN_PAGE = 4;
