+++ src/utils/product.ts
@@ -1,5 +1,5 @@
 import { getCategoryProducts } from '../api/Client';
-import { CENTS_PER_EURO } from '../constants/common';
+import { CENTS_PER_EURO, POPULAR_PRODUCTS_IN_PAGE } from '../constants/common';
 import { CategoryProduct } from '../types/types';
 
 export const covertPrice = (price: number): string => {
@@ -33,4 +33,12 @@ export const getPagesArray = (totalPages: number): number[] => {
     result.push(i + 1);
   }
   return result;
+};
+
+export const indexRandomProducts = (num: number): number => {
+  const index = Math.floor(Math.random() * num) - POPULAR_PRODUCTS_IN_PAGE;
+  if (index >= 0) {
+    return index;
+  }
+  return 0;
 };
